TASK: Simplified EventStore schema
This mainly renames the table columns to be more consistent and in sync with the rest of the framework:

    stream_name => stream
    commit_version => version
    event => payload
    recorded_at => recordedat

Furthermore this adds an auto-incremented primary key "id" as replacement for "event_version" and uses
VARCHARS where applicable for better performance.

Lastly this uses a regular DATETIME type for the "recordetat" timestamp. For more precision the metadata
currently still stores microseconds.
As a result we can get rid of the custom `DateTimeType`.

Resolves: #43