BUGFIX: Persist projection changes right away
When a projection is updated multiple times during one
request - for example, because an event listener triggers
a second command resulting in an event - a projection
might work with inconsistent data, because the update
caused by the first event hasn't been persisted yet.

This will always be the case if the projector
uses a finder to retrieve a read model which just has
been created during this request.

Therefore, read models are now persisted right away
when a projector calls `add`, `update` or `remove`.