FEATURE: Garbage Collection for Doctrine Projections
This adds a simple garbage collection to the `DoctrineProjectionPersistenceManager`
to avoid it from running into memory excess.

Background:
Especially when replaying projections there might be a lot of pending changes
in the Unit Of Work. This change adds a simple counter that persists pending changes
as soon as they reach a certain number (100).

In the long run this should probably be replaced by a configurable strategy.

Related: #47