BUGFIX: Don't store "version" in the EventStream
Previously the `EventStream` had a version, but this only
makes sense for aggregate streams.
For different partitioned streams (e.g. for projections or
other event listeners) this version is misleading because it
just contains the highest version within the many streams.

Note:

We could replace the version with the last *sequence number*
but currently that's not needed and without that we are still
free to implement actual streams that stay open.