FEATURE: Unique event identifiers
This consistently renames "id" to "sequenceNumber" where applicable
and adds a new property "identifier" to `WritableEvent` and `RawEvent`
that will be set to a UUID during event publishing.

This allows for correlation and de-duplication of events in respective
handlers.

Note:
With this change it's not yet possible to *set* the event identifier from
the application.

Related: #6