FEATURE: Catch up for asynchronous event listeners
This change introduces a new catch up command for
asynchronous event listeners (including process
managers) and a row of related features and changes.

The new CLI command is located in the EventStore
command controller (`eventstore:catchup`) and
basically works like the `projection:catchup`
command except that you can't specify a specific
event listener.

The log about applied events is stored in a new
generic "applied events log repository". This
repository replaces the specialized ProjectionState
repository used by the `AbstractDoctrineProjector`.
It is not mandatory to use this new repository,
but very convenient if you don't want to implement
your own storage.

Finally, there are a few clean ups and new methods
in the Event Publisher which hopefully make it a
bit clearer how the respective methods work.