FEATURE: Improve event:* commands for automation
This change improves the output of the `event:watch` and
`event:catchup` commands for debugging and automation.

By default the `watch` command will result in a row of
dots, one per cycle, which communicate that a catchup
run was executed, but no new events were applied. When
a new event was applied, an asterisk ("*") is displayed.

When the command is run with the `--verbose` option,
more explanation, including the event listener class
name and applied event type, is displayed.

As previously, `--quiet` will result in no output
whatsoever, except for errors.