FEATURE: Improved schema creation and connection status
This change improves the CLI commands for schema creation and
displaying the connection status.

The `eventstore:createschema` command will now check if the given
event stream table already exists and if it does, displays a friendly
message instead of failing impolitely.

The `evenstore:status` command will now display which tables is
configured for the event stream and if that table exists.