!!! TASK: Rework AbstractDoctrineProjection for composibility & explicitness
This change slightly flattens the inheritance hierarchy of the doctrine
projectors and makes interaction with their state more explicit.

This is a breaking change because it affects the public API of the
`AbstractDoctrineProjector` and subclasses.
The public methods `get()`, `add()`, `update()` and `remove()` have
been moved to a sub object `state`.
Adjusting the code is straight-forward:

Previously:

```
final class SomeProjector extends AbstractDoctrineProjector
{
    public function whenSomeEvent(SomeEvent $event)
    {
        $this->add(new Object());
        // ...
```

Now:

```
final class SomeProjector extends AbstractDoctrineProjector
{
    public function whenSomeEvent(SomeEvent $event)
    {
        $this->state->add(new Object());
        // ...

Related: #136