FEATURE: Support for multiple Event Store backends
This feature allows multiple Event Store backends to be used within the same Flow installation.
With that it's possible to use separate database tables or completely different backends for certain event types.

When using the provided `DoctrineEventStorage` backend, the database connection options is now merged with the Flows default persistence backend options.
As a result the `default` Event Store uses the same DB connection circumventing #42.

In order to store events from a package `Some.Package` to a different database table, that's all you need to configure:

```yaml
Neos:
  EventSourcing:
    EventStore:
      stores:
        'someCustomStore':
          boundedContexts:
            'Some.Package': true
          storage: 'Neos\EventSourcing\EventStore\Storage\Doctrine\DoctrineEventStorage'
          storageOptions:
            eventTableName: 'some_custom_event_table'
```

To use a separate database altogether, the `backendOptions` can be changed, too:

```yaml
Neos:
  EventSourcing:
    EventStore:
      stores:
        'someCustomStore':
          # ...
          storageOptions:
            backendOptions:
              driver: 'pdo_sqlite'
              path: '/some/path/events.db'
              username: ~
              password: ~
```

Or, of course, a completely different adapter implementation can be used, by setting the `storage` option to a class name that implements `EventStorageInterface`.

This change also overhauls the `eventstore` CLI replacing the `createSchema` and `dropSchema` commands with the more general purpose commands `setup`, `setupall` and `status`.

Resolves: #55
Fixes: #42