FEATURE: Introduce createQuery for Doctrine Projector
This introduces a `createQuery()` method for the `AbstractDoctrineProjector`
which allows for finding entries in event listeners of the concrete
projector.

Note that, especially during a projection replay, earlier changes might
not yet be persisted by Doctrine and thus can't be found by a query.
You will need to persist changes manually with `$this->projectionPersistenceManager->persistAll();`
in these cases.

Closes #136