BUGFIX: Fix EntityMananger initialization
When using the new `eventstore::setup*` command(s) introduced
with #138 in a Neos installation an exception was thrown:

    Unknown column type "flow_json_array" requested.

Reason for this is the `EntityManagerConfiguration::enhanceEntityManager`
slot that is only invoked if the `EntityManager` is acutally used.

This patch fixes that by injecting the Doctrine `ObjectManager` to
the `ConnectionFactory` so that the signal is fired.

Related: #55