FEATURE: Events with metadata
This change provides a simple and straight-forward way to specify metadata
along with an event when it is published.

In order to specify the metadata, simply wrap the event to be published with
and `EventWithMetadata` object and set the metadata there. This wrapper object
can be published like a normal even using `publish()` or `publishMany()`.