FEATURE: Allow using a unified Repository for all Aggregates
This change adds the Repository method `loadByTypeAndIdentifier` which returns an aggregate of the given class and identifier. This way it is no longer necessary to create `{FooBar}Repository` classes that simply extend this class and are otherwise empty and instead allow to just inject a generic AggregateRepository in all CommandHandlers.

This generic class will be provided in a follow-up once an agreement on this behaviour and the method name has been found.