FEATURE: Add notion of correlation and causation Id
This is a new approach to introduce the notion of optional
correlation/causation ids to events.

Usage:

    $eventWithCorrelationId = new EventWithMetadata($originalEvent, ['correlationId' => $id]);
    $eventWithCausationId = new EventWithMetadata($originalEvent, ['causationId' => $id]);

This also reworks the `EventStreamFilterInterface` to be easier to extend/implement
and adds a new implementation to filter for the `correlationId` (as preparation for event-sourced
Process Managers).

Breaking Change:

This is a breaking change because it changes the `EventStreamFilterInterface`.
Besides, the schema of the doctrine based event-store is updated, adding the
two columns `correlationid` and `causationid`.
To update the db accordingly, just run

    ./flow eventstore:setupall

again.

Resolves: #6