FEATURE: Allow fetching of single events
This adds a new method `EventStore::getOne()` that allows
for fetching single events matching a given filter.

Besids this adds a new filter `StreamNameAndVersion` that
addresses a single event in a given stream.

Background:
This is the foundation to a more flexible & stable event handling
that enables the isolated processing of events in a separate
process.

Resolves: #164