BUGFIX: fix Event Decorators to properly work when they are nested
Before this change, when using multiple event decorators, the system
did not work properly. As an example, imagine you had the following
structure:

EventWithMetadata
  - EventWithIdentifier
    - Event

This led to the situation that the EventWithIdentifier was not used,
because EventWithMetadata directly unwrapped the EventWithIdentifier,
effectively discarding the identifier. The same problem happens vice-
versa with EventWithMetadata.

Now, we properly fix this by always storing the full event hierarchy;
and having a set of utility methods to extract the Identifiers, Metadata,
and Raw Events.