BUGFIX: Prevent UniqueConstraintViolationException when committing events
Introduces commit retries with exponential backoff in order to avoid
`UniqueConstraintViolationException` from being thrown when many events
are published almost at the same time from different processes.

Background:

Even though retrieving the currently highest version number for a stream and
the insert of new event rows is done transactionally that didn't prevent the
same max version from being retrieved by a 2nd process.
Fortunately we have a unique key constraint on the db level to prevent inconsistent
data, but the exception wasn't nice so this introduces some retry mechanism.

Note: After the maximum number of attempts are exceeded (currently eight) this
will still fail and throw a `ConcurrencyException` in that case.
Currently this will only happen after 1.2 seconds so it is quite unlikely.

I failed to cover the fixed behavior with meaningful tests (lacking the 2nd process
to actually provoke the bug) but I tested it manually by executing the following
code from two separate CLI commands:

```php
$streamName = StreamName::fromString('test');
$events = [];
for ($i = 0; $i < 10000; $i ++) {
    $events[] = new SomethingHasHappened($i);
}
$this->eventStore->commit($streamName, DomainEvents::fromArray($events));
```

The result is:

* 20.000 events in the database
* no gaps in the stream version
* a single gap of 1 in the sequence number

The gap in the sequence number arises from the parallel writes and can't
be avoided AFAIK, but they should not pose a problem anyways.

Fixes: #245