BUGFIX: Relax DefaultEventToListenerMappingProvider checks
With #256 the `DefaultEventToListenerMappingProvider` was introduced
with a very defensive behavior that prevents event store configurations
without matching Listeners.

This was taking it a little far and creates a hen-egg situation for the
initial setup because it requires an `EventListenerInterface` instance
to exist before a store can be configured.

With this change it's allowed to leave out the `listeners` configuration:

```yaml
Neos:
  EventSourcing:
    EventStore:
      stores:
        'event-store':
          storage: 'Neos\EventSourcing\EventStore\Storage\Doctrine\DoctrineEventStorage'
```

If one is specified it is still validated of course, so the follwing example will lead
to an exception if no EventListener exists yet:

```yaml
Neos:
  EventSourcing:
    EventStore:
      stores:
        'event-store':
          storage: 'Neos\EventSourcing\EventStore\Storage\Doctrine\DoctrineEventStorage'
          listeners:
            '.*': true
```