FEATURE: Introduce projection:catchup command
This change introduces a new CLI command which allows for applying
new events for a given or all projections. It's also possible to
specify the sequence number of an event up to which events should be
applied.

This command is meant to be used for debugging and other rather un-
common needs. In general, events should be applied by other means,
for example through the `CatchUpEventListenerJob`.

Resolves #270