Use class name instead of "self"
Makes sure that the actual class name is used instead of the "self"
keyword: If the class is ever proxied (and even final class can be
proxied), "self" would refer to the proxy, not to the original class.

It's safe to use the original class name since that will stay intact
even if the class is proxied.

#309